+++ graph/graph-application/src/main/kotlin/eu/tib/orkg/prototype/statements/api/resource-commands.kt
@@ -40,6 +40,7 @@ interface UpdateResourceUseCase {
         val classes: Set<ThingId>? = null,
         val observatoryId: ObservatoryId? = null,
         val organizationId: OrganizationId? = null,
+        val extractionMethod: ExtractionMethod? = null
     )
 }